requirements.txt
google-generativeai
streamlit
python-dotenv
tensorflow
transformers
langchain
